requirements.txt
Django==3.0
djongo==1.3.3
python-dotenv==0.20.0
pymongo==3.11.1
sqlparse==0.2.4

